package name is singular
