+++ src/main/java/kr/pincoin/durian/notification/controller/AligoController.java
@@ -1,10 +1,10 @@
-package kr.pincoin.durian.notifications.controller;
+package kr.pincoin.durian.notification.controller;
 
 import jakarta.validation.Valid;
 import kr.pincoin.durian.common.exception.ApiException;
-import kr.pincoin.durian.notifications.controller.dto.AligoSendRequest;
-import kr.pincoin.durian.notifications.controller.dto.AligoSendResponse;
-import kr.pincoin.durian.notifications.service.AligoService;
+import kr.pincoin.durian.notification.controller.dto.AligoSendRequest;
+import kr.pincoin.durian.notification.controller.dto.AligoSendResponse;
+import kr.pincoin.durian.notification.service.AligoService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
