+++ src/main/java/kr/pincoin/durian/notification/controller/LineNotifyController.java
@@ -1,10 +1,10 @@
-package kr.pincoin.durian.notifications.controller;
+package kr.pincoin.durian.notification.controller;
 
 import jakarta.validation.Valid;
 import kr.pincoin.durian.common.exception.ApiException;
-import kr.pincoin.durian.notifications.controller.dto.LineNotifyRequest;
-import kr.pincoin.durian.notifications.controller.dto.LineNotifyResponse;
-import kr.pincoin.durian.notifications.service.LineNotifyService;
+import kr.pincoin.durian.notification.controller.dto.LineNotifyRequest;
+import kr.pincoin.durian.notification.controller.dto.LineNotifyResponse;
+import kr.pincoin.durian.notification.service.LineNotifyService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
