+++ src/main/java/kr/pincoin/durian/notification/controller/dto/AligoSendResponse.java
@@ -1,4 +1,4 @@
-package kr.pincoin.durian.notifications.service.dto;
+package kr.pincoin.durian.notification.controller.dto;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -8,40 +8,35 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class AligoSendResult {
-    @JsonProperty("result_code")
+public class AligoSendResponse {
+    @JsonProperty("resultCode")
     private String resultCode;
 
     @JsonProperty("message")
     private String message;
 
-    @JsonProperty("msg_id")
+    @JsonProperty("msgId")
     @JsonInclude(JsonInclude.Include.NON_NULL)
     private String msgId;
 
-    @JsonProperty("success_cnt")
+    @JsonProperty("successCnt")
     @JsonInclude(JsonInclude.Include.NON_NULL)
     private Long successCnt;
 
-    @JsonProperty("error_cnt")
+    @JsonProperty("errorCnt")
     @JsonInclude(JsonInclude.Include.NON_NULL)
     private Long errorCnt;
 
-    @JsonProperty("msg_type")
+    @JsonProperty("msgType")
     @JsonInclude(JsonInclude.Include.NON_NULL)
     private String msgType;
 
-    public AligoSendResult(String resultCode, String message) {
-        this.resultCode = resultCode;
-        this.message = message;
-    }
-
-    public AligoSendResult(String resultCode,
-                           String message,
-                           String msgId,
-                           Long successCnt,
-                           Long errorCnt,
-                           String msgType) {
+    public AligoSendResponse(String resultCode,
+                             String message,
+                             String msgId,
+                             Long successCnt,
+                             Long errorCnt,
+                             String msgType) {
         this.resultCode = resultCode;
         this.message = message;
         this.msgId = msgId;
