+++ src/main/java/kr/pincoin/durian/notification/controller/dto/LineNotifyRequest.java
@@ -1,4 +1,4 @@
-package kr.pincoin.durian.notifications.controller.dto;
+package kr.pincoin.durian.notification.controller.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AccessLevel;
@@ -7,10 +7,7 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class AligoSendRequest {
-    @JsonProperty("phone")
-    private String phone;
-
+public class LineNotifyRequest {
     @JsonProperty("message")
     private String message;
 }
