+++ src/main/java/kr/pincoin/durian/notification/service/dto/AligoSendResult.java
@@ -1,4 +1,4 @@
-package kr.pincoin.durian.notifications.controller.dto;
+package kr.pincoin.durian.notification.service.dto;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -8,35 +8,40 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class AligoSendResponse {
-    @JsonProperty("resultCode")
+public class AligoSendResult {
+    @JsonProperty("result_code")
     private String resultCode;
 
     @JsonProperty("message")
     private String message;
 
-    @JsonProperty("msgId")
+    @JsonProperty("msg_id")
     @JsonInclude(JsonInclude.Include.NON_NULL)
     private String msgId;
 
-    @JsonProperty("successCnt")
+    @JsonProperty("success_cnt")
     @JsonInclude(JsonInclude.Include.NON_NULL)
     private Long successCnt;
 
-    @JsonProperty("errorCnt")
+    @JsonProperty("error_cnt")
     @JsonInclude(JsonInclude.Include.NON_NULL)
     private Long errorCnt;
 
-    @JsonProperty("msgType")
+    @JsonProperty("msg_type")
     @JsonInclude(JsonInclude.Include.NON_NULL)
     private String msgType;
 
-    public AligoSendResponse(String resultCode,
-                             String message,
-                             String msgId,
-                             Long successCnt,
-                             Long errorCnt,
-                             String msgType) {
+    public AligoSendResult(String resultCode, String message) {
+        this.resultCode = resultCode;
+        this.message = message;
+    }
+
+    public AligoSendResult(String resultCode,
+                           String message,
+                           String msgId,
+                           Long successCnt,
+                           Long errorCnt,
+                           String msgType) {
         this.resultCode = resultCode;
         this.message = message;
         this.msgId = msgId;
