+++ src/main/java/kr/pincoin/durian/notification/service/dto/MailgunSendResult.java
@@ -1,4 +1,4 @@
-package kr.pincoin.durian.notifications.controller.dto;
+package kr.pincoin.durian.notification.service.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AccessLevel;
@@ -7,15 +7,14 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class MailgunSendResponse {
+public class MailgunSendResult {
     @JsonProperty("id")
     private String id;
 
     @JsonProperty("message")
     private String message;
 
-    public MailgunSendResponse(String id,
-                               String message) {
+    public MailgunSendResult(String id, String message) {
         this.id = id;
         this.message = message;
     }
